Use UINT32 for certain indices
(cherry picked from commit 4a6ce392ab801f660a2da6a1f6c02cb2b0fd39a5)
(cherry picked from commit d9b3d908b6a82472d00f0162428b1663b1bf1291)
(cherry picked from commit b12e274af6dd875755047264bcaf4c59e6dc1d82)
(cherry picked from commit 90bb3bfc14a5bfb6f8f8ceacd8a2919e1ee6dd54)

+++ js/symbol/symbol_instances.js
@@ -19,10 +19,10 @@ const SymbolInstancesArray = createStructArrayType({
         { type: 'Uint16', name: 'textBoxEndIndex' },
         { type: 'Uint16', name: 'iconBoxStartIndex' },
         { type: 'Uint16', name: 'iconBoxEndIndex' },
-        { type: 'Uint16', name: 'glyphQuadStartIndex' },
-        { type: 'Uint16', name: 'glyphQuadEndIndex' },
-        { type: 'Uint16', name: 'iconQuadStartIndex' },
-        { type: 'Uint16', name: 'iconQuadEndIndex' },
+        { type: 'Uint32', name: 'glyphQuadStartIndex' },
+        { type: 'Uint32', name: 'glyphQuadEndIndex' },
+        { type: 'Uint32', name: 'iconQuadStartIndex' },
+        { type: 'Uint32', name: 'iconQuadEndIndex' },
 
         // each symbolInstance is centered around the anchor point
         { type: 'Int16', name: 'anchorPointX' },
